ICAforECGrecordings

Documentation for ICAforECGrecordings.

ICAforECGrecordings.cumulant_matrices

ICAforECGrecordings.jade

ICAforECGrecordings.joint_diagonalization

ICAforECGrecordings.load_example_data

ICAforECGrecordings.plot_dataset

ICAforECGrecordings.read_dataset_from_dat

ICAforECGrecordings.shibbs

ICAforECGrecordings.solve

ICAforECGrecordings.solve

ICAforECGrecordings.whiten

ICAforECGrecordings.cumulant_matrices — Method

cumulant_matrices(X::AbstractMatrix, m::Number)

Blind separation of real signals with SHIBBS.

Arguments

X::AbstractMatrix: Matrix that contains the signals that have to be unmixed.

m::Number: The number of signals that should be extracted from X.

Returns CM=cumulant_matrices(X, m) a NxN*nbcm cumulant matrix.

ICAforECGrecordings.jade — Method

jade(data::AbstractMatrix{<:Real}) -> Matrix{Float64}

Perform Blind Source Separation on multichannel time-series data using the JADE (Joint Approximate Diagonalization of Eigen-matrices) algorithm.

Arguments

data_w::AbstractMatrix{<:Real} A matrix of size (n_samples, n_signals + 1) where the first column is time in seconds and the remaining columns are the observed signals.

W_white::AbstractMatrix{<:Real} The whitening matrix of size (n_signals, n_signals) used to preprocess the data.

Returns

A matrix of size (n_samples, n_signals + 1) where the first column is time and the remaining columns are the estimated source signals after applying JADE.

ICAforECGrecordings.joint_diagonalization — Method

joint_diagonalization(CM::AbstractMatrix, threshhold::Real, m::Number)

Blind separation of real signals with SHIBBS.

Arguments

CM::AbstractMatrix: Matrix that contains the signals that have to be unmixed.

-threshhold::Real: A threshhold for the for the rotation

m::Number: The number of signals that should be extracted from CM.

Returns V = joint_diagonalization(CM, threshhold, m) a NxN diagonalized matrix.

ICAforECGrecordings.load_example_data — Method

load_example_data()

Load example data from a .dat file for testing purposes. Returns a matrix containing the data.

ICAforECGrecordings.plot_dataset — Method

plot_dataset(data::AbstractMatrix{Float64})

Plot a dataset with time on the x-axis and multiple signals on the y-axis.

Arguments

data::AbstractMatrix{Float64}: Matrix where the first column represents time and the remaining columns represent different signals.

ICAforECGrecordings.read_dataset_from_dat — Method

read_dataset_from_dat(Path :: String)

Reads a dataset from a .dat file.

Arguments

Path::String: The path to the .dat file.

Returns

A matrix of type Float64 containing the data read from the file.

ICAforECGrecordings.shibbs — Function

shibbs(X::AbstractMatrix,m::Int)

Blind separation of real signals with SHIBBS.

Arguments

X::AbstractMatrix: Matrix that contains the signals that have to be unmixed.

m=nothing: amount of signals that should be extracted from X. default same amount of singals as source

Returns B=shibbsR(X) a mxn separating matrix such that S=B*X is an mxT matrix of estimated source signals.

ICAforECGrecordings.solve — Method

solve(seperator::AbstractSeperator, data::AbstractMatrix) -> AbstractMatrix

Performs Independent Component Analysis (ICA) using the specified algorithm on the provided data.

Arguments

seperator: An instance of AbstractSeperator, which can be JadeSeperator, ShibbsSeperator, or PicardoSeperator.

data: A matrix of size (n_{samples}, n_{signals} + 1) where the first column is time in seconds and data[:, 2:end] contains the signal measurements.

Returns

A matrix of size (n_{samples}, n_{signals} + 1) where the first column is time and [:, 2:end] are the separated signals.

ICAforECGrecordings.solve — Method

solve(seperator::ShibbsSeperator, data::AbstractMatrix) -> AbstractMatrix

Performs Independent Component Analysis (ICA) using the Shibbs algorithm on the provided data.

Arguments

seperator: An instance of ShibbsSeperator.

data: A matrix of size (n_{samples}, n_{signals} + 1) where the first column is time in seconds and data[:, 2:end] contains the signal measurements.

Returns

A matrix of size (n_{samples}, n_{signals} + 1) where the first column is time and [:, 2:end] are the separated signals.

ICAforECGrecordings.whiten — Method

whiten(data::AbstractMatrix) -> Tuple{Matrix, Matrix}

Performs whitening on the given data matrix.

Whitening transforms the input signals so that each signal has unit variance and all signals are uncorrelated. This is a common preprocessing step before applying Independent Component Analysis (ICA).

Arguments

data: An abstract matrix of size (n_samples, n_signals+1) where the first column is time and the remaining n_signals columns are the observed signals.

Returns

data_white::Matrix: A matrix of the same size as data, where the first column is the original time vector and the remaining columns are the whitened signals.

W_white::Matrix: The whitening transform matrix of size (n_signals, n_signals) such that X_white = X_centered * W_white.